Move plssvm::csvm implementation to cpp file where necessary. Re-add const where possible.
